refactor: code & tests refactoring
