+++ src/Andy.Cli/Services/TextWrapping/KnuthPlassTextWrapper.cs
@@ -1,8 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Text.RegularExpressions;
 
 namespace Andy.Cli.Services.TextWrapping;
 
@@ -10,16 +6,14 @@ namespace Andy.Cli.Services.TextWrapping;
 /// Advanced text wrapper implementing the Knuth-Plass line breaking algorithm.
 /// This provides optimal line breaking with minimal "badness" across the entire paragraph.
 /// </summary>
-public class KnuthPlassTextWrapper : ITextWrapper
+public class KnuthPlassTextWrapper : TextWrapperBase
 {
-    private readonly IHyphenationService _hyphenationService;
-
     public KnuthPlassTextWrapper(IHyphenationService hyphenationService)
+        : base(hyphenationService)
     {
-        _hyphenationService = hyphenationService ?? throw new ArgumentNullException(nameof(hyphenationService));
     }
 
-    public WrappedText WrapText(string text, int maxWidth, TextWrappingOptions? options = default)
+    public override WrappedText WrapText(string text, int maxWidth, TextWrappingOptions? options = default)
     {
         if (string.IsNullOrEmpty(text) || maxWidth <= 0)
             return new WrappedText(new[] { text ?? string.Empty });
@@ -53,15 +47,6 @@ public class KnuthPlassTextWrapper : ITextWrapper
         return new WrappedText(lines, hasHyphenation);
     }
 
-    public int MeasureLineCount(string text, int maxWidth, TextWrappingOptions? options = default)
-    {
-        if (string.IsNullOrEmpty(text) || maxWidth <= 0)
-            return 1;
-
-        var wrappedText = WrapText(text, maxWidth, options);
-        return wrappedText.LineCount;
-    }
-
     private (List<string> lines, bool hasHyphenation) WrapWithKnuthPlass(string line, int maxWidth, TextWrappingOptions options)
     {
         // For now, implement a simplified version that respects maxWidth
@@ -82,7 +67,7 @@ public class KnuthPlassTextWrapper : ITextWrapper
             if (currentWidth + spaceWidth + wordWidth > maxWidth)
             {
                 // Try to hyphenate the word to fit on current line if hyphenation is enabled
-                if (currentLine.Length > 0 && options.EnableHyphenation && _hyphenationService.CanHyphenate(word))
+                if (currentLine.Length > 0 && options.EnableHyphenation && HyphenationService.CanHyphenate(word))
                 {
                     var remainingWidth = maxWidth - currentWidth - spaceWidth;
                     var hyphenated = TryHyphenateToFit(word, remainingWidth, maxWidth, options);
@@ -164,145 +149,15 @@ public class KnuthPlassTextWrapper : ITextWrapper
         return (wrappedLines, hasHyphenation);
     }
 
-    private List<string> TokenizeIntoWords(string text)
-    {
-        // Simple word tokenization - split on whitespace
-        var words = new List<string>();
-        var currentWord = new StringBuilder();
-
-        foreach (char c in text)
-        {
-            if (char.IsWhiteSpace(c))
-            {
-                if (currentWord.Length > 0)
-                {
-                    words.Add(currentWord.ToString());
-                    currentWord.Clear();
-                }
-                // Skip whitespace in simple mode
-            }
-            else
-            {
-                currentWord.Append(c);
-            }
-        }
-
-        if (currentWord.Length > 0)
-        {
-            words.Add(currentWord.ToString());
-        }
-
-        return words;
-    }
-
-    private (List<string> lines, bool hasHyphenation) BreakLongWord(string word, int firstSegmentWidth, int fullWidth, TextWrappingOptions options)
-    {
-        var brokenLines = new List<string>();
-        bool hasHyphenation = false;
-
-        if (!options.EnableHyphenation || !_hyphenationService.CanHyphenate(word))
-        {
-            // No hyphenation - just break at character boundaries
-            // First segment uses firstSegmentWidth, subsequent use fullWidth
-            int pos = 0;
-            while (pos < word.Length)
-            {
-                int segmentWidth = pos == 0 ? firstSegmentWidth : fullWidth;
-                // Don't use negative or zero widths - ensure at least 1
-                if (segmentWidth <= 0)
-                {
-                    segmentWidth = 1;
-                }
-                // Ensure segment doesn't exceed maxWidth
-                segmentWidth = Math.Min(segmentWidth, fullWidth);
-                var segment = word.Substring(pos, Math.Min(segmentWidth, word.Length - pos));
-                brokenLines.Add(segment);
-                pos += segment.Length;
-            }
-            return (brokenLines, hasHyphenation);
-        }
-
-        // Try to hyphenate
-        var hyphenationPoints = _hyphenationService.GetHyphenationPoints(word);
-        var currentPos = 0;
-
-        while (currentPos < word.Length)
-        {
-            // For first segment, use remaining width; for subsequent segments, use full width
-            var remainingWidth = currentPos == 0 ? firstSegmentWidth : fullWidth;
-            var bestBreak = -1;
-
-            // Find the best hyphenation point within the remaining width
-            // Choose the latest possible point to maximize line width usage
-            // Account for hyphen character (+1) when checking fit
-            foreach (var point in hyphenationPoints)
-            {
-                var segmentLength = point - currentPos;
-                var segmentWithHyphenLength = segmentLength + 1; // Account for hyphen
-                
-                if (point > currentPos && segmentWithHyphenLength <= remainingWidth)
-                {
-                    // Check if this break point respects minimum fragment lengths
-                    var leftLength = segmentLength;
-                    
-                    // Only check left fragment minimum length - allow any length to use full width
-                    if (leftLength >= options.MinHyphenationLength &&
-                        point > bestBreak) // Prefer later break points
-                    {
-                        bestBreak = point;
-                    }
-                }
-            }
-
-            if (bestBreak >= 0)
-            {
-                // Use hyphenation - check if adding hyphen would exceed width
-                var segmentLength = bestBreak - currentPos;
-                var segmentWithHyphenLength = segmentLength + 1; // Account for hyphen
-                
-                // Check if segment with hyphen fits
-                if (segmentWithHyphenLength <= remainingWidth)
-                {
-                    // Use hyphenation
-                    var segment = word.Substring(currentPos, segmentLength) + "-";
-                    brokenLines.Add(segment);
-                    currentPos = bestBreak;
-                    hasHyphenation = true;
-                }
-                else
-                {
-                    // Hyphen would exceed width, fall back to character breaking
-                    var segmentLength2 = Math.Max(1, Math.Min(remainingWidth, word.Length - currentPos));
-                    segmentLength2 = Math.Min(segmentLength2, fullWidth);
-                    var segment = word.Substring(currentPos, segmentLength2);
-                    brokenLines.Add(segment);
-                    currentPos += segmentLength2;
-                }
-            }
-            else
-            {
-                // Fall back to character boundary breaking
-                var segmentLength = Math.Max(1, Math.Min(remainingWidth, word.Length - currentPos));
-                // Ensure segment doesn't exceed maxWidth
-                segmentLength = Math.Min(segmentLength, fullWidth);
-                var segment = word.Substring(currentPos, segmentLength);
-                brokenLines.Add(segment);
-                currentPos += segmentLength;
-            }
-        }
-
-        return (brokenLines, hasHyphenation);
-    }
-
     private HyphenationResult TryHyphenateToFit(string word, int availableWidth, int maxWidth, TextWrappingOptions options)
     {
-        if (!options.EnableHyphenation || !_hyphenationService.CanHyphenate(word) || availableWidth < options.MinHyphenationLength + 1)
+        if (!options.EnableHyphenation || !HyphenationService.CanHyphenate(word) || availableWidth < options.MinHyphenationLength + 1)
         {
             // Need at least min length + 1 for hyphen
             return new HyphenationResult { Success = false };
         }
 
-        var hyphenationPoints = _hyphenationService.GetHyphenationPoints(word);
+        var hyphenationPoints = HyphenationService.GetHyphenationPoints(word);
         int bestBreak = -1;
         int bestLeftLength = 0;
 
@@ -348,18 +203,6 @@ public class KnuthPlassTextWrapper : ITextWrapper
         }
 
         return new HyphenationResult { Success = false };
-    }
-
-    private int GetDisplayWidth(string text)
-    {
-        // For now, assume 1 character = 1 display width
-        // This could be enhanced to handle full-width characters, etc.
-        return text.Length;
-    }
-
-    private static bool IsVowel(char c)
-    {
-        return "aeiouAEIOU".IndexOf(c) >= 0;
     }
 
     private class HyphenationResult
