+++ src/Andy.Cli/Services/TextWrapping/TextWrappingOptionsExtensions.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-
 namespace Andy.Cli.Services.TextWrapping;
 
 /// <summary>
