+++ tests/Andy.Cli.Tests/Services/MarkdownContentProcessorTests.cs
@@ -1,5 +1,4 @@
 using Andy.Cli.Services.ContentPipeline;
-using Xunit;
 
 namespace Andy.Cli.Tests.Services;
